multi_tool_agent/requirements.txt
google-adk
google-generativeai>=0.3.0
python-dotenv>=1.0.0
